실습4 uint to int 변경

+++ Lab4/Lab4/MultiSet.cs
@@ -24,11 +24,11 @@ namespace Lab4
             return false;
         }
 
-        public uint GetMultiplicity(string element)
+        public int GetMultiplicity(string element)
         {
             if (mSet.Contains(element))
             {
-                return (uint)mSet.FindAll(s => s.Equals(element)).Count;
+                return mSet.FindAll(s => s.Equals(element)).Count;
             }
 
             return 0;
@@ -47,10 +47,10 @@ namespace Lab4
 
             foreach (var element in union)
             {
-                uint a = this.GetMultiplicity(element);
-                uint b = other.GetMultiplicity(element);
+                int a = this.GetMultiplicity(element);
+                int b = other.GetMultiplicity(element);
 
-                uint count = Math.Max(a, b);
+                int count = Math.Max(a, b);
 
                 while (count > 0)
                 {
@@ -70,10 +70,10 @@ namespace Lab4
 
             foreach (var element in union)
             {
-                uint a = this.GetMultiplicity(element);
-                uint b = other.GetMultiplicity(element);
+                int a = this.GetMultiplicity(element);
+                int b = other.GetMultiplicity(element);
 
-                uint count = Math.Min(a, b);
+                int count = Math.Min(a, b);
 
                 while (count > 0)
                 {
@@ -93,10 +93,10 @@ namespace Lab4
 
             foreach (var element in union)
             {
-                uint a = this.GetMultiplicity(element);
-                uint b = other.GetMultiplicity(element);
+                int a = this.GetMultiplicity(element);
+                int b = other.GetMultiplicity(element);
 
-                uint count = Math.Max(a - b, 0);
+                int count = Math.Max(a - b, 0);
 
                 while (count > 0)
                 {
